requirements.txt
Flask
flask-cors
flask-pymongo
pymongo
marshmallow
requests
google-genai
dotenv-load_dotenv
bcrypt
PyJWT
